SCAN4NET-407 Support JreProvisioningTest SQC ITs on Linux & MacOS
