+++ templates/unix-qa-stage.yml
@@ -40,6 +40,10 @@ stages:
               PRODUCT: "SonarQube"
               SQ_VERSION: "LATEST_RELEASE"
               TEST_INCLUDE: "**/sonarqube/ScannerTest*,**/sonarqube/SslTest*,**/sonarqube/JreProvisioningTest*"
+            Cloud:
+              PRODUCT: "SonarCloud"
+              SQ_VERSION: ""
+              TEST_INCLUDE: "**/sonarcloud/JreProvisioningTest*"
             Others:
               PRODUCT: "others"
               SQ_VERSION: ""
